% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colors.R
\name{agree_disagree_pal}
\alias{agree_disagree_pal}
\title{A muted, qualitative color palette}
\usage{
agree_disagree_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(agree_disagree_pal()(4))
}
