% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colors.R
\name{idealpoints_pal}
\alias{idealpoints_pal}
\title{A muted, qualitative color palette}
\usage{
idealpoints_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(idealpoints_pal()(4))
}
