% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota.R
\name{import_roboto}
\alias{import_roboto}
\title{Import roboto fonts
Make sure the roboto fonts are in your path}
\usage{
import_roboto()
}
\description{
Import roboto fonts
Make sure the roboto fonts are in your path
}
\examples{
\dontrun{
import_roboto()
}
}
\concept{fonts}
