% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colors.R
\name{jota_pal}
\alias{jota_pal}
\title{A muted, qualitative color palette}
\usage{
jota_pal()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(jota_pal()(4))
}
