% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Normalization}
\alias{Normalization}
\title{Normalization
Normalization of a numeric vector}
\usage{
Normalization(x)
}
\arguments{
\item{x}{The numeric vector}
}
\description{
Normalization
Normalization of a numeric vector
}
