% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{real_format}
\alias{real_format}
\title{Format as Brazilian Real
Format numbers as Brazilian Reais}
\usage{
real_format(x)
}
\arguments{
\item{x}{The vector}
}
\description{
Format as Brazilian Real
Format numbers as Brazilian Reais
}
