% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota.R
\docType{data}
\name{font_rb}
\alias{font_rb}
\title{Roboto Regular font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_rb
}
\description{
\code{font_rb} == "\verb{Roboto Bold}"
}
\keyword{datasets}
