% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota.R
\docType{data}
\name{font_rr}
\alias{font_rr}
\title{Roboto Regular font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_rr
}
\description{
\code{font_rr} == "\verb{Roboto Regular}"
}
\keyword{datasets}
