% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota.R
\docType{data}
\name{font_sr}
\alias{font_sr}
\title{Sans Regular font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_sr
}
\description{
\code{font_sr} == "\verb{Sans Regular}"
}
\keyword{datasets}
