% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_ggplot.R
\name{save_ggplot}
\alias{save_ggplot}
\title{Save ggplot2 objetc}
\usage{
save_ggplot(plot, width, height, save_filepath)
}
\description{
A wrapper function for the ggplot2 ggsave with custom parameters.
}
