% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_horizontal.R
\name{theme_horizontal}
\alias{theme_horizontal}
\title{A [ggplot2] theme with opinionated aesthetics for <https://jota.info>}
\usage{
theme_horizontal(base_size = 10, base_family = "sans",
  horizontal = TRUE, dark_panel = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{horizontal}{\code{logical} horizontal axis lines?}

\item{dark_panel}{\code{logical} darker background for panel?}
}
\description{
A [ggplot2] theme with opinionated aesthetics for <https://jota.info>
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)
data(Governismo)

ggplot(Governismo, aes(x=D1, y=D2)) +
  geom_point() +
  labs(x="IDEOLOGIA (D1)", y="POSICIONAMENTO ECONÔMICO (D2)",
       title="Um exemplo de scatterplot",
       subtitle="Apoio ao texto da Reforma: Threshold 80\% a 20\%",
      caption="Fonte: https://jota.info") +
  theme_horizontal()
 }

}
\seealso{
Other themes: \code{\link{theme_jota}}
}
\concept{themes}
