% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota.R
\name{theme_jota}
\alias{theme_jota}
\title{A \link{ggplot2} theme with opinionated aesthetics for \url{https://jota.info}}
\usage{
theme_jota(base_family = "Roboto Regular", base_size = 11.5,
  plot_title_family = base_family, plot_title_size = 18,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = if (.Platform$OS.type == "windows") "Roboto Regular"
  else "Arial", subtitle_size = 13, subtitle_face = "plain",
  subtitle_margin = 15, strip_text_family = base_family,
  strip_text_size = 12, strip_text_face = "plain",
  caption_family = if (.Platform$OS.type == "windows") "Roboto Regular"
  else "Arial", caption_size = 9, caption_face = "plain",
  caption_margin = 10, axis_text_size = base_size,
  axis_title_family = base_family, axis_title_size = 9,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30), grid_col = "#9e9c90",
  grid = TRUE, axis_col = "#232323", axis = FALSE, ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of \code{[blmcrt]}}

\item{plot_margin}{plot margin (specify with \link[ggplot2:margin]{ggplot2::margin})}

\item{grid_col}{grid color}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis_col}{axis color}

\item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{ticks if \code{TRUE} add ticks}
}
\description{
This theme uses 'Roboto Regular' as the default typoghraphy and dashed line for grid plot.
}
\details{
Notice: You should \url{import_roboto} first and also install the fonts on your
system before trying to use this theme.

There is an option \code{rJOTA.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\section{Why Roboto?}{

It's free and has tolerable kerning pairs and multiple weights.
}

\examples{
\dontrun{
library(ggplot2)
library(dplyr)
data(Governismo)

ggplot(Governismo, aes(x=D1, y=D2)) +
  geom_point() +
  labs(x="IDEOLOGIA (D1)", y="POSICIONAMENTO ECONÔMICO (D2)"
       title="Um exemplo de scatterplot",
       subtitle="Apoio ao texto da Reforma: Threshold 80\% a 20\%",
      caption="Fonte: https://jota.info") +
  theme_jota()


# seminal bar chart

group_by(Governismo, Reeleito) \%>\%
 summarize(Indice = mean(Indice, na.rm=TRUE)) \%>\%
 mutate(Reeleito = ifelse(Reeleito==1, "Reeleito", "Novato")) \%>\%
 ggplot(aes(x=Reeleito, y=Indice)) +
 geom_col() +
 geom_text(aes(label=round(100*Indice,1)), nudge_y=.02) +
 labs(x="Status do parlamentar (Reeleitos vs Novatos)", y="Governismo (Indice)",
      title="Um exemplo de gráfico de barras",
      subtitle="Novatos compõem a principal força de apoio ao governo",
      caption="Fonte: https://jota.info") +
 theme_jota(grid="Y") +
 theme(axis.text.y=element_blank())
}

}
