% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jota_map.R
\name{theme_jota_map}
\alias{theme_jota_map}
\title{A \link{ggplot2} theme with opinionated aesthetics for diplaying maps at \url{https://jota.info}}
\usage{
theme_jota_map(...)
}
\arguments{
\item{...}{map theme parameters}
}
\description{
This theme uses 'Roboto Regular' as the default typoghraphy and white panel plot.
}
\details{
Notice: You should \url{import_roboto} first and also install the fonts on your
system before trying to use this theme.

There is an option \code{jotaR.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\section{Why Roboto?}{

It's free and has tolerable kerning pairs and multiple weights. It's also the default font type used at JOTA materials.
}

