% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Clean theme for maps}
\usage{
theme_map(base_size = 9, base_family = "sans")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A clean theme that is good for displaying maps from
\code{\link[ggplot2]{geom_map}()}.
}
\seealso{
Other themes: \code{\link{theme_horizontal}},
  \code{\link{theme_jota}}
}
\concept{themes}
