% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{virgula_format}
\alias{virgula_format}
\title{Format number with comma
Format numbers with comma}
\usage{
virgula_format(x)
}
\arguments{
\item{x}{The vector}
}
\description{
Format number with comma
Format numbers with comma
}
